tidy up main.py and move init methods to its own helper file
